+++ include/kfr/audio.hpp
@@ -27,6 +27,7 @@
 #include "audio/data.hpp"
 #include "audio/decoder.hpp"
 #include "audio/encoder.hpp"
+#include "audio/filter.hpp"
 #include "audio/io.hpp"
 
 namespace kfr